Add ToJSON ICECandidate method
With this change we can always exchange ICECandidateInit when signaling
