+++ go.mod
@@ -11,7 +11,7 @@ require (
 	github.com/pion/rtcp v1.2.1
 	github.com/pion/rtp v1.1.3
 	github.com/pion/sctp v1.6.4
-	github.com/pion/sdp/v2 v2.2.0
+	github.com/pion/sdp/v2 v2.3.0
 	github.com/pion/srtp v1.2.6
 	github.com/pion/transport v0.8.6
 	github.com/stretchr/testify v1.3.0
